+++ bitcoind-rpc/src/Bitcoin/Core/RPC/Generating.hs
@@ -2,7 +2,6 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
-{-# LANGUAGE TypeOperators #-}
 
 module Bitcoin.Core.RPC.Generating (
     generateToAddress,