+++ bitcoind-regtest/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 0.3.0.0 -- 2021-??-??
 
-* Expose bitcoind version in node handle
+* Adds logic to generate blocks with trannsactions (for fee estimation) and an
+  exec to run this logic
+* Adds a function to start up a cluster of connected nodes
+* Changes to `NodeHandle` (incl. adding bitcoind version)
 * Enable wallets 
 
 ## 0.2.0.0 -- 2020-11-01